torch -> jax img manipulation

+++ utils.py
@@ -95,7 +95,7 @@ def pre_process_pose(input_video, apply_pose_detect: bool = True):
         detected_map = cv2.resize(detected_map, (W, H), interpolation=cv2.INTER_NEAREST)
         detected_maps.append(detected_map[None])
     detected_maps = np.concatenate(detected_maps)
-    control = (detected_maps.copy()).float() / 255.0
+    control = (detected_maps.copy()) / 255.0
     return control #rearrange(control, 'f h w c -> f c h w')
